+++ rust/Cargo.toml
@@ -16,6 +16,5 @@ members = [
 	"qrcode_static",
 	"signer",
 	"transaction_parsing",
-	"transaction_signing",
-	"extrinsic_proof"
+	"transaction_signing"
 ]
