--- rust/extrinsic_proof/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "extrinsic_proof"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bitvec = "1.0.1"
-array-bytes = { version = "6.2.2", default-features = false }
-hex = "0.4.3"
-parser = {path = "../parser", default-features = false}
-printing_balance = {path = "../printing_balance"}
-blake3 = { version = "1.5.0", default-features = false }
-num-bigint = "0.4.4"
-codec = { package = "parity-scale-codec", version = "3.6.9", features = [ "derive", "bit-vec"] }
-sp-core = {git = "https://github.com/paritytech/substrate", rev = "49734dd1d72a00b9d3b87ba397661a63e0e17af3", default-features = false, features = ["full_crypto"]}
-sp-runtime = {git = "https://github.com/paritytech/substrate", rev = "49734dd1d72a00b9d3b87ba397661a63e0e17af3", default-features = false}
-scale-decode = { version = "0.13.0", default-features = false }
-scale-info = { version = "2.10.0", default-features = false }
-
-[dev-dependencies]
-pretty_assertions = "1"
-
-[lib]
-name = "extrinsic_proof"
-crate-type = ["lib"]
