+++ rust/parser/Cargo.toml
@@ -15,11 +15,13 @@ lazy_static = "1.4.0"
 num-bigint = "0.4.4"
 parity-scale-codec = {version = "3.6.9", features = ["derive", "bit-vec"]}
 printing_balance = {path = "../printing_balance"}
+blake3 = { version = "1.5.0", default-features = false }
 regex = "1.10.3"
 scale-info = "2.10.0"
 sp-arithmetic = {git = "https://github.com/paritytech/substrate", rev = "49734dd1d72a00b9d3b87ba397661a63e0e17af3"}
 sp-core = {git = "https://github.com/paritytech/substrate", rev = "49734dd1d72a00b9d3b87ba397661a63e0e17af3", default-features = false, features = ["full_crypto"]}
 sp-runtime = {git = "https://github.com/paritytech/substrate", rev = "49734dd1d72a00b9d3b87ba397661a63e0e17af3", default-features = false}
+scale-decode = { version = "0.13.0", default-features = false }
 thiserror = "1.0.57"
 
 [dev-dependencies]
