+++ rust/parser/src/tests.rs
@@ -1,9 +1,15 @@
 use std::convert::TryInto;
+use std::io::Read;
 use std::str::FromStr;
 
 use crate::method::OlderMeta;
 use crate::Error;
-use crate::{parse_set, MetadataBundle};
+use crate::{
+    parse_set, 
+    MetadataBundle,
+    types:: MetadataProof,
+    decoding_with_proof::{decode_call, decode_extensions},
+};
 use defaults::default_types_vec;
 use definitions::metadata::info_from_metadata;
 use definitions::network_specs::ShortSpecs;
@@ -566,4 +572,43 @@ tx_version: 2,
 block_hash: 2470dff6295dd9bb3e5a89c9eb7647d7c5ae525618d77757171718dc034be8f5"#;
 
     assert_eq!(reply, reply_known)
+}
+
+#[test]
+fn parse_proof() {
+    let data = hex::decode("6c0c1c73705f636f72651c65643235353139245369676e6174757265000400168d0101205b75383b2036345d8901000340000000038d0100034100000003f501082873705f72756e74696d65384d756c74695369676e6174757265011c4564323535313904001689010148656432353531393a3a5369676e617475726500b502082873705f72756e74696d65384d756c74695369676e6174757265011c53723235353139040016b9020148737232353531393a3a5369676e617475726504b502082873705f72756e74696d65384d756c74695369676e617475726501144563647361040016bd02014065636473613a3a5369676e617475726508b5020c1c73705f636f72651c73723235353139245369676e6174757265000400168d0101205b75383b2036345db9020c1c73705f636f7265146563647361245369676e617475726500040016f501017c5b75383b205349474e41545552455f53455249414c495a45445f53495a455dbd0204184f7074696f6e0110536f6d650400160400040d0710306672616d655f73797374656d28657874656e73696f6e733c636865636b5f6d6f7274616c69747938436865636b4d6f7274616c697479000400161507010c4572611107102873705f72756e74696d651c67656e657269630c6572610c4572610120496d6d6f7274616c0000150710306672616d655f73797374656d28657874656e73696f6e732c636865636b5f6e6f6e636528436865636b4e6f6e6365000400110120543a3a4e6f6e63651907086870616c6c65745f7472616e73616374696f6e5f7061796d656e74604368617267655472616e73616374696f6e5061796d656e7400040013013042616c616e63654f663c543e1d0708746672616d655f6d657461646174615f686173685f657874656e73696f6e44436865636b4d6574616461746148617368000401106d6f646516250701104d6f6465210708746672616d655f6d657461646174615f686173685f657874656e73696f6e104d6f6465011c456e61626c6564000425070c1c73705f636f72651863727970746f2c4163636f756e7449643332000400160401205b75383b2033325d000003200000000304083c7072696d69746976655f74797065731048323536000400160401205b75383b2033325d0c000203100003140000000358085873746167696e675f6b7573616d615f72756e74696d652c52756e74696d6543616c6c012042616c616e636573040016090101b50173656c663a3a73705f6170695f68696464656e5f696e636c756465735f636f6e7374727563745f72756e74696d653a3a68696464656e5f696e636c7564653a3a64697370617463680a3a3a43616c6c61626c6543616c6c466f723c42616c616e6365732c2052756e74696d653e10c40c2873705f72756e74696d65306d756c746961646472657373304d756c746941646472657373010849640400160001244163636f756e7449640005010c2873705f72756e74696d65306d756c746961646472657373304d756c7469416464726573730114496e64657804001501304163636f756e74496e6465780405010c2873705f72756e74696d65306d756c746961646472657373304d756c746941646472657373010c52617704001610011c5665633c75383e0805010c2873705f72756e74696d65306d756c746961646472657373304d756c74694164647265737301244164647265737333320400160401205b75383b2033325d0c05010c2873705f72756e74696d65306d756c746961646472657373304d756c74694164647265737301244164647265737332300400165801205b75383b2032305d1005010c3c70616c6c65745f62616c616e6365731870616c6c65741043616c6c01307472616e736665725f616c6c0801106465737416050101504163636f756e7449644c6f6f6b75704f663c543e01286b6565705f616c697665000110626f6f6c1009016c2f08000030080000a20800001c0900001d0900001e0900001f09000020090000b40c0000b50c0000b60c0000b60d0000b70d0000b80d0000ba0d0000dd060000de060000e0060000e10600001807000085070000d6070000d7070000d8070000d9070000da070000de070000fc8bde9d188f3499593ba7f99020bd66dec949f4be9dc6a66758ca866422245347ef6d44700f24f683d8feaf3f4d6edc6a3db690b577810dd93f5238bcfa124a13ac779fbf4c947aec99bb2634eb9d29b6d52fc5963772873a83d48577c7214722ce1976678609c6f557300ff3e56ffcc3defa3f6fbf4a57d28f1e253a571616ccd3f8c3eb9a50cab7215b568eb10ad38608c9a0d411d033ec7bf89d55f4dcb998d30740deb50756e990d6fd819078ba15f0abd766b0324919ebae3f4c395f27cf78377a1a4bd823be5173718f60efc6a4e538be88615c9cec8f15d735e5a058f8c2d730d4ac8ce7ebe47ea17e27fa9276172aead98e82107983689bd41a939a57ab751746a8f56503041dfefa81c2574de5ded59abfcb9f0c0a6f65855ad8545be8da772d99c79aa5b36243bce64d8072de79fd0cc35315d4e1fcec75d465e67601c6e51472477a25c7a925567050140b94afaebf680c302ca89fa4659a6dda70b658d6b273a2c386d92b18641cf47a6059babf7c9355a33dddd60b75f7c041f3b1356d74be287ea67f1dc6c69103f75ac1ad80f0fbdd3b75b8f7638609a506af911f7bb7dde1e7262aecfe0cb03e2c28babcf1d7bd9d6bd6525c081ae26bf95af18a99d1b11dd75919ae80006dda74e6ae72bdbe61810623705e293ea27ca920fbb1731995be8d202457a897111c7420acfe2e36b6c58c050b51285b163d46131f39e2a5d54d0778f6283e57c308fdf9b93f57ba191dca9e6fb2d3c8cf62de9ed7f7d75fbceacdc662468b66b8e9f5ec7310387353707436b38118f01b12463e5054ee57e6f738ef49d67799a81d0f68c158ff1e472d9316fc7a7a1300a70c1a6e071beac89f4e0287b5e275903441da86396bc9327cc646f253dae1dc8910d0bea3b25cb532f2c5e17a776d0420ef599a40fdbe3000545fe2240558f96c5dea9d24cf20971575d638657a6567f9acddb073f07008e116213faf2d2eb064da536d452e54b19cc160fadaebd431cbebf3fa5bf463955c738c7d498841b0f9f7dfcca5ebe84049282f197cb5313538e8d85be64d3adb490407ae4a4b2d0baad9deaf9a70548731d77ecaf9943a189451af961b16d1b0d781f0889c04a46e30968b02d7f40078cf1c61d2f2744148612dfb7971b6e5c9ba87f0894b4ef4695534e5069af0b15360134700bd7a90625116f96ffbdc75b8dbe6a06de8eb15467c7ed5a538db02f4ca0eece9b49e78998b7422f9861fb8ee3ed316f9da97227605fc5baa7eb3a33be3682bfbe0b899d842f8bfd9d8015c7a20a1b9309556c52b05f66f27567eeb322ead2601f7b7315d1268054b905aa828678da59c62c8449e1e44050328edb8fdd8afe843668a86405c8c1c8747592a8a4065fbaf7c3e437da250a43f58499f077c3fd8f28a2c2b4051eb32388bf1551155402d102014148766f79fa308349211bba1a514131dbca953340657c94a9ec09eecf2229a5c3f1bbf43ae86064a8fd921000809ca3967d9aa5df5d57f3db593158b0b45c383a7eb404ed4839189a897142e4ef45c76d48ff137140991f84b5560750c35bdebdca525645a843858ca14f9eaeae5bf6c660803d7382e9bd5d20df02a301a3fa474de97e86fbfad84e58bb1a65e1dace12466fe5ae1ae8f6823613d76837b57ab006f143fc6692ba0cc1f77fdc2eeb4beb73680d96a9c79f6762b7db1d2c31729d365f469e6d3814e151905ab6499c76ccb290d3cc9fa2f41bd00139140b7206e31ff9d98e4beef1d1a2e20bbc4a225773b9327590503729e589420b1a75d86eaabff22d7ceb0a3bce7985ac4ce86e36c57c1716dc9c7e103b2ad5a424a8d0e86dfa17f9a071a4ca92b98305c4a6a097cf2c7aeb6a138e586e83ec9dc850e880454b5533277669bc8f535ef347749341984137b61bccb4b97bc8fcc0576ea2276932f69f3ffaff017312e602d04cb25dacd91c8f9b1797c535f550349c8d621ec291121cf9ac3879b870b158d60694f4bb8e4ffa15642b7b34f08e4e193d3948771a858cfe628f6b448ce2a5c2dfe378662956fc81bcc55181f9b46402fae502ced038b37582eb85ba6f478bd380a5ae1a8c6ee4d58aefae37bad6545830a19d35ad57d054db47da4b720e2c8ddd8d5ae0e4cb2bc2899380e6bca62a25762aae935d615b47d59006f6ab8b70fd6eea47df9996f36b035b06ca1d530a27b1a390399118ece662370974a1bcc221a9724d6cc7c99f16c3e0e07f029e68b5b1de3abafc70bc66303bc9e89e521487c83729c7255aa1eb141a6ddc50a1604f42382c05d3d3e3ad4f8666381447f04e7c663c2c181773b65ca9c4003bafc1c5ddc30fe8ce4e2c99f9d4b7a8eb23f9fc207c9088f7f3fe1cc725a648850f9474e94b36eabaecc1a0b991007bb2a1fe0eed54bbcecfa39ccad91f790f4bbf93db3b21c088b8f65a27103a04abd67e8436744eb4ee6d03fc727eb6357a0e6137602cfd4cc55ba904e1287d1a641aad217883231ad3e30bcb6bbc69becb38ee680a7086a3fb43ea507a7f903cb14477122b7ee2feb89906de66339c66c9801f8ea1f41afbfad78ba52820787af4f0288dd3a73f790d0ed3ccb97bc34140c59638d895e4ee67d8f262a521826e71f5225900365e413f63a1d2852e9f20cd96f7b422435a736619ae698c90665fb75f49ad043a3ccdf1e2d06656f498c736a9f8ca46535b191b54d8c5985ad83b32be5993f83ec1fa50a74f414480e7ab146bdeb34f8685241fa21a7da2d1641a8dc8dd533905ace593f2ea4f6399062d06639f7630ecb1cf3f5f6db95948568eb7535f3cc0b184a00cb355bae78e95dfdda04f41226dc1908f40fbaf55b0416050116c416b5022448436865636b4e6f6e5a65726f53656e646572151540436865636b5370656356657273696f6e150538436865636b547856657273696f6e150530436865636b47656e6573697315160c38436865636b4d6f7274616c697479161107160c28436865636b4e6f6e6365161907152c436865636b5765696768741515604368617267655472616e73616374696f6e5061796d656e74161d071544436865636b4d6574616461746148617368162107160d07154a0f00186b7573616d6102000c0c4b534d").unwrap();
+    let metadata = MetadataProof::decode(&mut &data[..]).ok().unwrap();
+    println!("{:?}", metadata);
+}
+
+// TODO: add more tests
+
+#[test]
+fn parse_transfer() {
+    let data = hex::decode("60000341000000031102082873705f72756e74696d65384d756c74695369676e6174757265011c4564323535313904001685010148656432353531393a3a5369676e617475726500f504082873705f72756e74696d65384d756c74695369676e6174757265011c5372323535313904001685010148737232353531393a3a5369676e617475726504f504082873705f72756e74696d65384d756c74695369676e6174757265011445636473610400161102014065636473613a3a5369676e617475726508f50404184f7074696f6e0110536f6d65040016040004990610306672616d655f73797374656d28657874656e73696f6e733c636865636b5f6d6f7274616c69747938436865636b4d6f7274616c69747900040016a106010c4572619d06102873705f72756e74696d651c67656e657269630c6572610c4572610120496d6d6f7274616c0000a10610306672616d655f73797374656d28657874656e73696f6e732c636865636b5f6e6f6e636528436865636b4e6f6e6365000400110120543a3a4e6f6e6365a506086870616c6c65745f7472616e73616374696f6e5f7061796d656e74604368617267655472616e73616374696f6e5061796d656e7400040013013042616c616e63654f663c543ea90608746672616d655f6d657461646174615f686173685f657874656e73696f6e44436865636b4d6574616461746148617368000401106d6f646516b10601104d6f6465ad0608746672616d655f6d657461646174615f686173685f657874656e73696f6e104d6f6465011c456e61626c65640004b1060c1c73705f636f72651863727970746f2c4163636f756e7449643332000400160401205b75383b2033325d000003200000000304083c7072696d69746976655f74797065731048323536000400160401205b75383b2033325d0c000203100003140000000350085873746167696e675f6b7573616d615f72756e74696d652c52756e74696d6543616c6c012042616c616e636573040016090101b50173656c663a3a73705f6170695f68696464656e5f696e636c756465735f636f6e7374727563745f72756e74696d653a3a68696464656e5f696e636c7564653a3a64697370617463680a3a3a43616c6c61626c6543616c6c466f723c42616c616e6365732c2052756e74696d653e10c40c2873705f72756e74696d65306d756c746961646472657373304d756c746941646472657373010849640400160001244163636f756e7449640005010c2873705f72756e74696d65306d756c746961646472657373304d756c7469416464726573730114496e64657804001501304163636f756e74496e6465780405010c2873705f72756e74696d65306d756c746961646472657373304d756c746941646472657373010c52617704001610011c5665633c75383e0805010c2873705f72756e74696d65306d756c746961646472657373304d756c74694164647265737301244164647265737333320400160401205b75383b2033325d0c05010c2873705f72756e74696d65306d756c746961646472657373304d756c74694164647265737301244164647265737332300400165001205b75383b2032305d1005010c3c70616c6c65745f62616c616e6365731870616c6c65741043616c6c01507472616e736665725f616c6c6f775f64656174680801106465737416050101504163636f756e7449644c6f6f6b75704f663c543e011476616c7565130128543a3a42616c616e63650009010003400000000385016069080000280a0000290a00002a0a0000260c0000270c0000280c0000280d0000290d00002a0d00002c0d00009606000097060000990600009a060000cf06000040070000900700009107000092070000930700009407000095070000e9070000e4703e8894f16ebdddd78bcbfb4515610caf771595904de66167f2e185bc20891b9ba7258b77cb883b4c1ad21b03819c4b5647a208d42701a72926e1f5c593f165b1b1519755cd94d82bf864dbdcb98d5d35a7b7efe61221fb5a9b2b477313a7f148123bc34b73c6412247df27dea44beb0c68d2e0842847a43fe256d2a4274421f6d13467e99bc29a73393436d4a1d5fb64d2386b785f36b5265195e27e4ae87d01d0606446a7c97d4f66cb2a9bd43bb9318bb189d15bcaa30678b84488d95c21c7009de256b20e8803393c9d8799abb62239d9d97fec8897a90ae7797e9ac8d27237d41c085f8838e3b938e2b24ea0bf5993b843e76331cfbd12381c936493f0454f3c8c24969855dd9e680d2c27952b7d64e2d369b0ad5750cea0ad4e237fccbd52da3c79c836bae74e0587c71c8c48d29f04cfe6afa00cbb3bdb1580351ab313b39b0f57aca514fec890f6304c6b5233e3a815802439cf19fe9ecec2c88334a693b63d584dc3581db368dc3be143bb376166d5d828bce0b2b03e158d9bbbec1a87211b6097dca8537d4f3ecc5fa0b20ddcd4a1d32e8028182b7e770809a5f565333c9be85eecbaefb0bffc8cedcffd54dbfb28640e0f0367ef26bdaac1148386f2efcd3bcd1ccd95e40d02a84c438e5791385f1448de9b93a1947b91c1ea5c4cb5a6c60f946bb39ddfa419cc5bd97478067c8a22f3f79729125d9362684b10053e17d18a5ff8171dbd25dbe0eb3e16eb66f22bf7bbb4f4d3bec8d8cc77414c957764ce39bd3babf1d4bda04d64d1df4b1fc2bd8c197181c961ac43585cfb8f0db0d0c98594d52e43b83beaaebe89857a7a23b5c9dbaec324bd53639ac8e1a6db3e4c6e73b7704b51fde8a42b4c78f49dceeb1a57286e0f7d25b4b55b1dbbc9e8d1036ac7f65d110f790f142ee609b1d6ddace5ed19f17833ac91458858b3d66fd5597864036218b09bbd058b1d56c6dd4c16d02c4127b3ffb2446a6f23fc76dc3eafe2f3a41a0a9e2b2486aa10dcd14a1ba0eb231799948264f33998119d039de987660ae1afb7fdb3c643d5cc67ac5105f74c2f8c3f61440562b2de579dd86ca136a91ed479244ca6a0cc3e7879e4c7b3844104d3ab9c7fdb6c8c30d20fe7c1507546f97650eedcfb67f99afc7086be9fdb4823d012a6872e992e22e7e3a12d2c40720ea9ed350daa484f4ecd0fc77cada99bd7731e280a5ad01e89875788a59612ffe1cfff0ac75e4b70b3116b535ab276d8cd561fd101e7b5600ffd41f0246ff387cb531313e556e93ff9901e0ee287c1bc610c5685c58a36653ba49926dcd0889f624dd6d1bf2a102b07ab1703f8646885efba024ba335422cceb78fbcd3814e151905ab6499c76ccb290d3cc9fa2f41bd00139140b7206e31ff9d98e46bcb4a390fb36de6ae704ea64833eaceb9a16e4ccdc6694265ab5b5bf9d3c3cc5902b4e9428cdf83c31e32e354e62ece41a061cf458f67e0f7b0f50de0d6e945f348206210d990d1edac69e5f4662956151b2627ba1d1780085e943297ec28acc959b5b61034ddbe0748d5b74e251bed7d6fc93fb5ea49d3a4d20a8377fbe21a0a6f9a77af07a6e327c294a37e08cb6ee843efc6d8d37e0bac411a39ddb82dde562e99e5de9a2ac842938216124e57f4bbb18942cb3168f9356c52a5c7c66ce0ecc709ea59127487c3062d51189d09fbdfbd1ac5bd003db0c5619eb211b58941a7fd45b12cbecf31add5b6dbb749115183dcc059860ffaceb2514015bc0536b96de7122697c842ea9016de8f549729f4b1b0b4807150e5204b459cf5a48b24745538e012aec0584d0c622ab61705b3d00dfaeaba7731c2a2f684f130a3fc8c60733b72288a9dd65cac4a6aa965d291acdee7f300153ef6b4466d936f5f90006939f6fdc2094d8b3e3ea6e9d27c35781422c6a0d91a5dd92f8551307151074260313a5328d3c0248944b03603f022a16ef789790494093e21bda021136dfa7fb64d504354eeb528654b065d7713311c94bf2f9a3519222cf260eb8c7f6ed8d64f96e69975bdf28f17003f59fd4cd0d2581d5826489d17a1f05e9e76979e2850d35bb0bdb9ba49e3d399dd0a6f9294a59731c0085017e74a1b6690e4ad0e2406a80665fb75f49ad043a3ccdf1e2d06656f498c736a9f8ca46535b191b54d8c5985925d75cbff17e16d959a0f9c03dd8f15204a2eb33c5bcfa9fc1b71dbc2cc99e24bb3cd887b4e0bd1df7da5a8ca5b908a24f18a982453e9ba50954886f68cfe97c15bd19b56f6f955a9c726c60a7dfd38a140befa8a6cdd05686ca2916b8d6e731fd783dbe50b4479b9e102fb7906d1512cda1fd495bcff19fc36788d2ad2dd6fd55d99d99c4e4a82990b3a738a6b1c9bd0efc790993fec6a4605deb295277e013ab6d381b57453f5a62fae2c1df4a5a4d71d7480f301e1effec910fb778c6ee40ea6bf3b3e5c325a6126523a0e42c8158d371951cb9587196450c0681696fc9503b3ba991b967965009e3d0dc5effd655616e2aeb0d0328799b161337164ae4a1e34f865e3346b96d22e21ecf9d4e26fae7237deb69234a031e5874285f320060416050116c416f5042448436865636b4e6f6e5a65726f53656e646572151540436865636b5370656356657273696f6e150538436865636b547856657273696f6e150530436865636b47656e6573697315160c38436865636b4d6f7274616c697479169d06160c28436865636b4e6f6e636516a506152c436865636b5765696768741515604368617267655472616e73616374696f6e5061796d656e7416a9061544436865636b4d657461646174614861736816ad06169906e0510f00186b7573616d6102000c0c4b534d").unwrap();
+    let metadata = MetadataProof::decode(&mut &data[..]).ok().unwrap();
+
+    println!("{:?}", metadata);
+
+    let call = hex::decode("040000e2e058da1316f8425be6c6f7104bb44c96fa29ea8b890b4c2866ba8bb6bc67b807007c118d35").unwrap();
+
+    let cards = decode_call(&mut &call[..], &metadata).unwrap();
+    println!("{:?}", cards);
+}
+
+#[test]
+fn parse_extensions() {
+    let data = hex::decode("60000341000000031102082873705f72756e74696d65384d756c74695369676e6174757265011c4564323535313904001685010148656432353531393a3a5369676e617475726500f504082873705f72756e74696d65384d756c74695369676e6174757265011c5372323535313904001685010148737232353531393a3a5369676e617475726504f504082873705f72756e74696d65384d756c74695369676e6174757265011445636473610400161102014065636473613a3a5369676e617475726508f50404184f7074696f6e0110536f6d65040016040004990610306672616d655f73797374656d28657874656e73696f6e733c636865636b5f6d6f7274616c69747938436865636b4d6f7274616c69747900040016a106010c4572619d06102873705f72756e74696d651c67656e657269630c6572610c4572610120496d6d6f7274616c0000a10610306672616d655f73797374656d28657874656e73696f6e732c636865636b5f6e6f6e636528436865636b4e6f6e6365000400110120543a3a4e6f6e6365a506086870616c6c65745f7472616e73616374696f6e5f7061796d656e74604368617267655472616e73616374696f6e5061796d656e7400040013013042616c616e63654f663c543ea90608746672616d655f6d657461646174615f686173685f657874656e73696f6e44436865636b4d6574616461746148617368000401106d6f646516b10601104d6f6465ad0608746672616d655f6d657461646174615f686173685f657874656e73696f6e104d6f6465011c456e61626c65640004b1060c1c73705f636f72651863727970746f2c4163636f756e7449643332000400160401205b75383b2033325d000003200000000304083c7072696d69746976655f74797065731048323536000400160401205b75383b2033325d0c000203100003140000000350085873746167696e675f6b7573616d615f72756e74696d652c52756e74696d6543616c6c012042616c616e636573040016090101b50173656c663a3a73705f6170695f68696464656e5f696e636c756465735f636f6e7374727563745f72756e74696d653a3a68696464656e5f696e636c7564653a3a64697370617463680a3a3a43616c6c61626c6543616c6c466f723c42616c616e6365732c2052756e74696d653e10c40c2873705f72756e74696d65306d756c746961646472657373304d756c746941646472657373010849640400160001244163636f756e7449640005010c2873705f72756e74696d65306d756c746961646472657373304d756c7469416464726573730114496e64657804001501304163636f756e74496e6465780405010c2873705f72756e74696d65306d756c746961646472657373304d756c746941646472657373010c52617704001610011c5665633c75383e0805010c2873705f72756e74696d65306d756c746961646472657373304d756c74694164647265737301244164647265737333320400160401205b75383b2033325d0c05010c2873705f72756e74696d65306d756c746961646472657373304d756c74694164647265737301244164647265737332300400165001205b75383b2032305d1005010c3c70616c6c65745f62616c616e6365731870616c6c65741043616c6c01507472616e736665725f616c6c6f775f64656174680801106465737416050101504163636f756e7449644c6f6f6b75704f663c543e011476616c7565130128543a3a42616c616e63650009010003400000000385016069080000280a0000290a00002a0a0000260c0000270c0000280c0000280d0000290d00002a0d00002c0d00009606000097060000990600009a060000cf06000040070000900700009107000092070000930700009407000095070000e9070000e4703e8894f16ebdddd78bcbfb4515610caf771595904de66167f2e185bc20891b9ba7258b77cb883b4c1ad21b03819c4b5647a208d42701a72926e1f5c593f165b1b1519755cd94d82bf864dbdcb98d5d35a7b7efe61221fb5a9b2b477313a7f148123bc34b73c6412247df27dea44beb0c68d2e0842847a43fe256d2a4274421f6d13467e99bc29a73393436d4a1d5fb64d2386b785f36b5265195e27e4ae87d01d0606446a7c97d4f66cb2a9bd43bb9318bb189d15bcaa30678b84488d95c21c7009de256b20e8803393c9d8799abb62239d9d97fec8897a90ae7797e9ac8d27237d41c085f8838e3b938e2b24ea0bf5993b843e76331cfbd12381c936493f0454f3c8c24969855dd9e680d2c27952b7d64e2d369b0ad5750cea0ad4e237fccbd52da3c79c836bae74e0587c71c8c48d29f04cfe6afa00cbb3bdb1580351ab313b39b0f57aca514fec890f6304c6b5233e3a815802439cf19fe9ecec2c88334a693b63d584dc3581db368dc3be143bb376166d5d828bce0b2b03e158d9bbbec1a87211b6097dca8537d4f3ecc5fa0b20ddcd4a1d32e8028182b7e770809a5f565333c9be85eecbaefb0bffc8cedcffd54dbfb28640e0f0367ef26bdaac1148386f2efcd3bcd1ccd95e40d02a84c438e5791385f1448de9b93a1947b91c1ea5c4cb5a6c60f946bb39ddfa419cc5bd97478067c8a22f3f79729125d9362684b10053e17d18a5ff8171dbd25dbe0eb3e16eb66f22bf7bbb4f4d3bec8d8cc77414c957764ce39bd3babf1d4bda04d64d1df4b1fc2bd8c197181c961ac43585cfb8f0db0d0c98594d52e43b83beaaebe89857a7a23b5c9dbaec324bd53639ac8e1a6db3e4c6e73b7704b51fde8a42b4c78f49dceeb1a57286e0f7d25b4b55b1dbbc9e8d1036ac7f65d110f790f142ee609b1d6ddace5ed19f17833ac91458858b3d66fd5597864036218b09bbd058b1d56c6dd4c16d02c4127b3ffb2446a6f23fc76dc3eafe2f3a41a0a9e2b2486aa10dcd14a1ba0eb231799948264f33998119d039de987660ae1afb7fdb3c643d5cc67ac5105f74c2f8c3f61440562b2de579dd86ca136a91ed479244ca6a0cc3e7879e4c7b3844104d3ab9c7fdb6c8c30d20fe7c1507546f97650eedcfb67f99afc7086be9fdb4823d012a6872e992e22e7e3a12d2c40720ea9ed350daa484f4ecd0fc77cada99bd7731e280a5ad01e89875788a59612ffe1cfff0ac75e4b70b3116b535ab276d8cd561fd101e7b5600ffd41f0246ff387cb531313e556e93ff9901e0ee287c1bc610c5685c58a36653ba49926dcd0889f624dd6d1bf2a102b07ab1703f8646885efba024ba335422cceb78fbcd3814e151905ab6499c76ccb290d3cc9fa2f41bd00139140b7206e31ff9d98e46bcb4a390fb36de6ae704ea64833eaceb9a16e4ccdc6694265ab5b5bf9d3c3cc5902b4e9428cdf83c31e32e354e62ece41a061cf458f67e0f7b0f50de0d6e945f348206210d990d1edac69e5f4662956151b2627ba1d1780085e943297ec28acc959b5b61034ddbe0748d5b74e251bed7d6fc93fb5ea49d3a4d20a8377fbe21a0a6f9a77af07a6e327c294a37e08cb6ee843efc6d8d37e0bac411a39ddb82dde562e99e5de9a2ac842938216124e57f4bbb18942cb3168f9356c52a5c7c66ce0ecc709ea59127487c3062d51189d09fbdfbd1ac5bd003db0c5619eb211b58941a7fd45b12cbecf31add5b6dbb749115183dcc059860ffaceb2514015bc0536b96de7122697c842ea9016de8f549729f4b1b0b4807150e5204b459cf5a48b24745538e012aec0584d0c622ab61705b3d00dfaeaba7731c2a2f684f130a3fc8c60733b72288a9dd65cac4a6aa965d291acdee7f300153ef6b4466d936f5f90006939f6fdc2094d8b3e3ea6e9d27c35781422c6a0d91a5dd92f8551307151074260313a5328d3c0248944b03603f022a16ef789790494093e21bda021136dfa7fb64d504354eeb528654b065d7713311c94bf2f9a3519222cf260eb8c7f6ed8d64f96e69975bdf28f17003f59fd4cd0d2581d5826489d17a1f05e9e76979e2850d35bb0bdb9ba49e3d399dd0a6f9294a59731c0085017e74a1b6690e4ad0e2406a80665fb75f49ad043a3ccdf1e2d06656f498c736a9f8ca46535b191b54d8c5985925d75cbff17e16d959a0f9c03dd8f15204a2eb33c5bcfa9fc1b71dbc2cc99e24bb3cd887b4e0bd1df7da5a8ca5b908a24f18a982453e9ba50954886f68cfe97c15bd19b56f6f955a9c726c60a7dfd38a140befa8a6cdd05686ca2916b8d6e731fd783dbe50b4479b9e102fb7906d1512cda1fd495bcff19fc36788d2ad2dd6fd55d99d99c4e4a82990b3a738a6b1c9bd0efc790993fec6a4605deb295277e013ab6d381b57453f5a62fae2c1df4a5a4d71d7480f301e1effec910fb778c6ee40ea6bf3b3e5c325a6126523a0e42c8158d371951cb9587196450c0681696fc9503b3ba991b967965009e3d0dc5effd655616e2aeb0d0328799b161337164ae4a1e34f865e3346b96d22e21ecf9d4e26fae7237deb69234a031e5874285f320060416050116c416f5042448436865636b4e6f6e5a65726f53656e646572151540436865636b5370656356657273696f6e150538436865636b547856657273696f6e150530436865636b47656e6573697315160c38436865636b4d6f7274616c697479169d06160c28436865636b4e6f6e636516a506152c436865636b5765696768741515604368617267655472616e73616374696f6e5061796d656e7416a9061544436865636b4d657461646174614861736816ad06169906e0510f00186b7573616d6102000c0c4b534d").unwrap();
+    let metadata = MetadataProof::decode(&mut &data[..]).ok().unwrap();
+
+    println!("{:?}", metadata);
+
+    let mut extension_data = hex::decode("00000001").unwrap();
+    let mut included_in_signature_data = hex::decode("e0510f001a000000b0a8d493285c2df73290dfb7e61f870f17b41801197a149ca93654499ea3dafeb0a8d493285c2df73290dfb7e61f870f17b41801197a149ca93654499ea3dafe0170d7bc8f2b306a20914d9dac518ccfd7b08e1e0404e0379534327560d5391c38").unwrap();
+
+    extension_data.append(&mut included_in_signature_data);
+
+    let cards = decode_extensions(&mut &extension_data[..], &metadata).unwrap();
+
+    println!("Extensions cards {:?}", cards);
 }
